refactor: reorganize ai enrichment tooling
