+++ pyproject.toml
@@ -1,0 +1,34 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "discos-analisis"
+version = "0.1.0"
+description = "Herramientas para enriquecer y analizar inventarios de discos duros."
+readme = "README.md"
+authors = [{ name = "MingoMedia" }]
+requires-python = ">=3.10"
+license = { text = "MIT" }
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3 :: Only",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = []
+
+[project.optional-dependencies]
+test = ["pytest", "flake8"]
+
+[project.scripts]
+discos-enrich = "discos_analisis.cli.enrich:main"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.package-data]
+"discos_analisis" = []
