+++ src/discos_analisis/__init__.py
@@ -1,0 +1,5 @@
+"""Herramientas de análisis y enriquecimiento para inventarios de discos."""
+
+from . import ai, annotations, inventory  # noqa: F401
+
+__all__ = ["ai", "annotations", "inventory"]
