+++ src/discos_analisis/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Comandos de línea de órdenes para discos_analisis."""
+
+from .enrich import main  # noqa: F401
+
+__all__ = ["main"]
